from_rows improve schema correctness
